requirements.txt
together
pandas
matplotlib
seaborn
openpyxl
python-docx
PyMuPDF
pytesseract
pillow
tabulate
streamlit  